Module warp_api_ffi::api::account

Account related API

Functions

delete_account

Delete an account

derive_keys

Derive keys using Zip-32

get_backup

Get the backup string. It is either the passphrase, the secret key or the viewing key depending on how the account was created

get_sk

Get the secret key. Returns empty string if the account has no secret key

get_taddr_balance

Retrieve the transparent balance from the LWD server

get_taddr_balance_default

Retrieve the transparent balance for the current account from the LWD server

import_from_zwl

Import a ZWL data file

import_sync_data

Import synchronization data obtained from external source

import_transparent_key

Update the transparent secret key for the given account from a derivation path

import_transparent_secret_key

Update the transparent secret key for the given account

new_account

Create a new account

new_diversified_address

Generate a new diversified address

new_sub_account

Create one or many sub accounts of the current account

reset_db

Reset the database

scan_transparent_accounts

Look for accounts that have some transparent balance. Stop when the gap limit is exceeded and no balance was found

truncate_data

Truncate all non account data for the current active coin

truncate_sync_data

Truncate all synchronization data for the current active coin