Module warp_api_ffi::api::sync

Warp Synchronize

Functions

coin_sync

Asynchronously perform warp sync

get_activation_date

Return the date of sapling activation

get_block_by_time

Return the block height for a given timestamp

get_latest_height

Return the latest block height

get_synced_height

Return the latest block height synchronized

rescan_from

Synchronize from a given height

rewind_to

Rewind to a previous block height

skip_to_last_height

Skip block synchronization and directly mark the chain synchronized Used for new accounts that have no transaction history